Add ability to cache text prop models
